use ajax instead of httpClient, restructure connection files
