--- src/connection/README.md
@@ -1,2 +1,0 @@
-[meredith-axios configs](https://github.com/whynotearth/meredith-axios/blob/master/dist/index.js)  
-[meredith-axios npm](https://www.npmjs.com/package/@whynotearth/meredith-axios)
