--- src/connection/init.js
@@ -1,3 +1,0 @@
-import { ajax } from './ajax.js';
-import { serviceOptions } from '@whynotearth/meredith-axios';
-serviceOptions.axios = ajax;
