+++ src/helpers.js
@@ -1,5 +1,3 @@
-import store from '@/store';
-
 export function randomId() {
   return Math.random()
     .toString()
@@ -49,12 +47,4 @@ export async function sleep(time) {
       resolve();
     }, time);
   });
-}
-
-export function getAuthHeaders() {
-  const JWTToken = store.getters['authKeep/token'];
-  if (!JWTToken) {
-    return;
-  }
-  return { Authorization: `Bearer ${JWTToken}` };
 }
