+++ src/main.js
@@ -1,6 +1,6 @@
 import Vue from 'vue';
 import App from './App.vue';
-import '@/connection/init';
+import '@/meredith-axios-init.js';
 import './registerServiceWorker';
 import router from './router';
 import store from './store';
