--- src/services/httpClient.js
@@ -1,6 +1,0 @@
-import axios from 'axios';
-
-export const httpClient = axios.create({
-  baseURL: process.env.VUE_APP_API_URL,
-  withCredentials: true
-});
